Units now save their LastState, which has all values from last Update call
Fixed a bug where you couldn't start moving horizontaly while jumping
Fixed a bug where the a unit would vibrate if his VelocityY was different than 30. It was due to a hardcoded value of 3 which was used to push the unit 3 pixels up until it came out of collision to find a valid position. The value is now 10% of VelocityY.
Midori's CollisionBox is now smaller and better fitted for her texture.